finsih

+++ app/Http/Controllers/JurusanController.php
@@ -87,6 +87,7 @@ class JurusanController extends Controller
     public function destroy(Jurusan $jurusan)
     {
         //
+        $this->authorize('delete',$jurusan);
         $jurusan->delete();
         return redirect('/')->with('pesan',"Jurusan $jurusan->nama_jurusan berhasil dihapus");
     }
